Moving the Mapping Health Check to its own package.
