+++ boggs/src/main/java/com/hartwig/healthchecks/boggs/adapter/BoggsAdapter.java
@@ -1,7 +1,7 @@
 package com.hartwig.healthchecks.boggs.adapter;
 
 import com.hartwig.healthchecks.boggs.flagstatreader.SambambaFlagStatParser;
-import com.hartwig.healthchecks.boggs.healthcheck.MappingHealthChecker;
+import com.hartwig.healthchecks.boggs.healthcheck.mapping.MappingHealthChecker;
 import com.hartwig.healthchecks.boggs.healthcheck.prestast.PrestastHealthChecker;
 import com.hartwig.healthchecks.boggs.healthcheck.prestast.PrestatsExtractor;
 import com.hartwig.healthchecks.boggs.io.PatientExtractor;
