+++ boggs/src/test/java/com/hartwig/healthchecks/boggs/healthchecker/MappingHealthCheckerTest.java
@@ -12,7 +12,7 @@ import com.hartwig.healthchecks.boggs.model.PatientData;
 import com.hartwig.healthchecks.boggs.model.SampleData;
 import com.hartwig.healthchecks.boggs.flagstatreader.FlagStatData;
 import com.hartwig.healthchecks.boggs.flagstatreader.FlagStatTestFactory;
-import com.hartwig.healthchecks.boggs.healthcheck.MappingHealthChecker;
+import com.hartwig.healthchecks.boggs.healthcheck.mapping.MappingHealthChecker;
 import com.hartwig.healthchecks.boggs.io.PatientExtractor;
 import com.hartwig.healthchecks.common.checks.HealthChecker;
